loro doc에서 f32 쓰던 부분 전부 u32/i32 로 변경 (#2994)

+++ crates/editor/src/transaction/root.rs
@@ -4,14 +4,14 @@ use crate::transaction::Transaction;
 use anyhow::Result;
 
 impl Transaction {
-    pub fn set_block_gap(&mut self, gap: f32) -> Result<bool> {
+    pub fn set_block_gap(&mut self, gap: u32) -> Result<bool> {
         let _ = self.doc().update_settings(|s| s.block_gap = gap);
         self.push_effect(Effect::SettingsChanged);
         self.push_effect(Effect::DocChanged);
         Ok(true)
     }
 
-    pub fn set_paragraph_indent(&mut self, indent: f32) -> Result<bool> {
+    pub fn set_paragraph_indent(&mut self, indent: u32) -> Result<bool> {
         let _ = self.doc().update_settings(|s| s.paragraph_indent = indent);
         self.push_effect(Effect::SettingsChanged);
         self.push_effect(Effect::DocChanged);
